refactor inputArrangers
