+++ src/index.ts
@@ -3,6 +3,7 @@ export * from './client.js'
 export * from './lib/config.js'
 export * from './lib/server-router-config.js'
 export * from './lib/server-router-config-defaults.js'
+export * from './lib/input-arrangers.js'
 export * from './lib/init-server-router-config.js'
 export * from './lib/common.js'
 export * from './lib/action-context.js'
